chore: move all builder to follow the same rules
