+++ test/reactions/github/sponsorship-created.spec.ts
@@ -1,4 +1,4 @@
-import { SponsorshipPayloadBuilder } from '../../builders/sponsorship-payload-builder';
+import { SponsorshipPayloadBuilder } from '../../builders/github/sponsorship-payload-builder';
 import { StreamLabsMock } from '../../__mocks__/StreamLabs';
 import { TwitchChatMock } from '../../__mocks__/TwitchChat';
 import { SponsorshipCreated } from '../../../src/reactions/github/sponsorship-created';
